# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FirmWizard API",
        "description": "AI-powered marketing platform for law firms",
        "contact": {
            "email": "support@firmwizard.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/onboarding/step/{step}": {
            "get": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Get specific wizard step",
                "description": "Show specific wizard step",
                "operationId": "1ffa9c8fa025e56616be9e51c4abff6c",
                "parameters": [
                    {
                        "name": "step",
                        "in": "path",
                        "description": "Step number (1-5)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display wizard step"
                    }
                }
            }
        }
    },
...
}